Add files via upload
A dataset has been converted into a time series and analyzed using the SARIMA model and the LSTM model, which is a subcategory of RNN. Forecasting has been performed using both methods, and their results have been compared. The project is written in Python. If you want to run the code you need to upload this dataset.